Feature/views

+++ app/controllers/actions_controller.rb
@@ -2,23 +2,35 @@
 class ActionsController < ApplicationController
   def new
     @action = Action.new
-    # authorize @action
+    authorize @action
+  end
+
+  def create
+    @action = Action.new
+    @action.assign_attributes(permitted_attributes(@action))
+
+    authorize @action
+
+    if @action.save
+      redirect_to case_path(@action), notice: 'Action created'
+    else
+      render :new
+    end
   end
 
   def index
-    # @actions = policy_scope(Action.all)
-    @actions = Action.all
-    # authorize @actions
+    @actions = policy_scope(Action.order(created_at: :desc).page(params[:page]))
+    authorize @actions
   end
 
   def edit
     @action = Action.find(params[:id])
-    # authorize @action
+    authorize @action
   end
 
   def update
     @action = Action.find(params[:id])
-    # authorize @action
+    authorize @action
 
     if @action.update_attributes(permitted_attributes(@action))
       redirect_to action_path, notice: 'Action updated'
@@ -29,12 +41,12 @@ class ActionsController < ApplicationController
 
   def show
     @action = Action.find(params[:id])
-    # authorize @action
+    authorize @action
   end
 
   def destroy
     @action = Action.find(params[:id])
-    # authorize @action
+    authorize @action
 
     if @action.destroy
       redirect_to actions_path, notice: 'Action deleted'
